Organize projecte files
